Module Distributed_lwt

A lwt based implementation of Distributed. Note, that this lwt based implmentation depends on the Logs_lwt library. In keeping with the usage conventions of the Logs library, this implementation does not define a source, it does not set the log level, and it does not define a reporter. The application is expected to define a source, set the log level, and define a reporter (see the examples).

author

essdotteedot <essdotteedot_at_gmail_dot_com>

version

0.2.0

module type CustomerLogger = sig ... end

This module provides a Log_lwt based logger to use.

module Make : functor (M : Distributed.Message_type) -> functor (L : CustomerLogger) -> Distributed.Process with type 'a io = 'a Lwt.t and type message_type = M.t

Functor to create a module of type Distributed.Process given a message module M of type Distributed.Message_type and a custom logger module L of type CustomerLogger.