Parameter Make.2-L

val msg : Logs.level -> 'a Logs_lwt.log

msg level logger returns a logger that can be use to log levels at the provided level. The returned thread only proceeds once the log operation is over.